src/main/java/student_pavel_sharkel/lesson_11/level_2/task_6_to_14/AndSearchCriteria.java
package student_pavel_sharkel.lesson_11.level_2.task_6_to_14;

class AndSearchCriteria implements SearchCriteria {

    private SearchCriteria leftCondition;
    private SearchCriteria rightCondition;

    public AndSearchCriteria(SearchCriteria leftCondition,
                             SearchCriteria rightCondition) {
        this.leftCondition = leftCondition;
        this.rightCondition = rightCondition;
    }

    public boolean match(Book book) {
        // допишите реализацию метода
        // return true - если книга удовлетворяет и левому и правому условию
        // иначе return false
        if (leftCondition.match(book) && rightCondition.match(book)) {
            return true;
        } else return false;
    }

}
